Improved Interface
Improved scripting interface.

Changed:
- solve now only does a solve, it will return an id number for the data
so it can be accessed later. This can now take a table with the solve
data instead of requiring a sepparate lua file.
- loadGeometry is now optional as it will take the geometry from the
SolverConfig, if it is the same then it will only load it once, saving
time/computing power.

Added:
- display this takes a solve id and displays the fields, it will also
return the number for the mode selected when the user closes the window.
This can be used to select modes.
- closestOverlap this takes two solves and a mode from one solve and
compares it to every mode in the second solve finding the closest
returning the mode and its overlap.
- outputField will output all the field components to a file for a given
solve and mode.
- getEffectiveIndex returns the effective index for a given solve and
mode.
- getAllEffectiveIndex returns a table with all the effective indices
for a given solve.